Exceptioni, Interface-i, loggeri, alertovi, novi entiteti, apstraktne klase i dovrsavanje starih todo zadataka.

+++ src/main/java/hr/java/player/util/DatotekaKorisnika.java
@@ -1,6 +1,11 @@
 package hr.java.player.util;
 
+import hr.java.player.entiteti.BaseUser;
+import hr.java.player.iznimke.DatotekaException;
+import javafx.scene.control.Alert;
+
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -15,7 +20,10 @@ import java.util.Map;
 public class DatotekaKorisnika {
     public static final String PATH_KORISNIKA = "dat/korisnici.txt";
 
-    public static Map<String,Integer> dohvatiSve(){
+    public static Map<String,Integer> dohvatiSve()throws DatotekaException{
+        if (!(new File(PATH_KORISNIKA).exists())){
+            throw new DatotekaException("Datoteka korisnika ne postoji!");
+        }
         Map<String,Integer> korisnici = new HashMap<>();
         try (BufferedReader in = new BufferedReader(new FileReader(PATH_KORISNIKA))) {
             String username;
@@ -24,20 +32,36 @@ public class DatotekaKorisnika {
                 passwordHash = Integer.parseInt(in.readLine());
                 korisnici.put(username,passwordHash);
             }
-        } catch (IOException ex) {
-            System.out.println(ex.getMessage());
+        } catch (IOException e) {
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Greska");
+            alert.setHeaderText("Greska prilikom citanja");
+            alert.setContentText("Dogodila se greska kod citanja sadrzaja datoteke.");
+            alert.showAndWait();
+            Logging.logger.error(e.getMessage(),e);
         }
         return korisnici;
     }
-    public static void unesiKorisnika(String username, Integer passwordHash){
+    public static void unesiKorisnika(BaseUser korisnik)throws DatotekaException{
+        if (!(new File(PATH_KORISNIKA).exists())){
+            throw new DatotekaException("Datoteka korisnika ne postoji!");
+        }
         try {
-            String korisnikString = username+'\n'+passwordHash+'\n';
+            String korisnikString = korisnik.username()+'\n'+korisnik.passwordHash()+'\n';
             Files.writeString(Path.of(PATH_KORISNIKA),korisnikString, StandardOpenOption.APPEND);
-        }catch (IOException ex){
-            ex.printStackTrace();
+        }catch (IOException e){
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Greska");
+            alert.setHeaderText("Greska prilikom citanja");
+            alert.setContentText("Dogodila se greska kod citanja sadrzaja datoteke.");
+            alert.showAndWait();
+            Logging.logger.error(e.getMessage(),e);
         }
     }
-    public static void obrisiKorisnika(String username){
+    public static void obrisiKorisnika(String username)throws DatotekaException{
+        if (!(new File(PATH_KORISNIKA).exists())){
+            throw new DatotekaException("Datoteka korisnika ne postoji!");
+        }
         try (BufferedReader in = new BufferedReader(new FileReader(PATH_KORISNIKA))) {
             String procitaniUsername;
             Path pathKorisnika = Path.of(PATH_KORISNIKA);
@@ -51,11 +75,19 @@ public class DatotekaKorisnika {
                 }
             }
             Files.write(pathKorisnika, procitaneLinije, StandardCharsets.UTF_8);
-        } catch (IOException ex) {
-            System.out.println(ex.getMessage());
+        } catch (IOException e) {
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Greska");
+            alert.setHeaderText("Greska prilikom citanja");
+            alert.setContentText("Dogodila se greska kod citanja sadrzaja datoteke.");
+            alert.showAndWait();
+            Logging.logger.error(e.getMessage(),e);
         }
     }
-    public static void promjeniUsername(String stariUsername, String noviUsername){
+    public static void promjeniUsername(String stariUsername, String noviUsername)throws DatotekaException{
+        if (!(new File(PATH_KORISNIKA).exists())){
+            throw new DatotekaException("Datoteka korisnika ne postoji!");
+        }
         try {
             Path pathKorisnika = Path.of(PATH_KORISNIKA);
             List<String> sveLinije = new ArrayList<>(Files.readAllLines(pathKorisnika));
@@ -67,22 +99,35 @@ public class DatotekaKorisnika {
             }
             Files.write(pathKorisnika, sveLinije, StandardCharsets.UTF_8);
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Greska");
+            alert.setHeaderText("Greska prilikom citanja");
+            alert.setContentText("Dogodila se greska kod citanja sadrzaja datoteke.");
+            alert.showAndWait();
+            Logging.logger.error(e.getMessage(),e);
         }
     }
-    public static void promjeniPassword(String username, Integer noviHash){
+    public static void promjeniPassword(BaseUser user)throws DatotekaException{
+        if (!(new File(PATH_KORISNIKA).exists())){
+            throw new DatotekaException("Datoteka korisnika ne postoji!");
+        }
         try {
             Path pathKorisnika = Path.of(PATH_KORISNIKA);
             List<String> sveLinije = new ArrayList<>(Files.readAllLines(pathKorisnika));
             for (int i=0;i<sveLinije.size();i++){
-                if (sveLinije.get(i).equals(username)){
-                    sveLinije.set(i+1,String.valueOf(noviHash));
+                if (sveLinije.get(i).equals(user.username())){
+                    sveLinije.set(i+1,String.valueOf(user.passwordHash()));
                     break;
                 }
             }
             Files.write(pathKorisnika, sveLinije, StandardCharsets.UTF_8);
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Greska");
+            alert.setHeaderText("Greska prilikom citanja");
+            alert.setContentText("Dogodila se greska kod citanja sadrzaja datoteke.");
+            alert.showAndWait();
+            Logging.logger.error(e.getMessage(),e);
         }
     }
 }
